ContentView.swift
Rather than using the removeAll() method and a for-loop to populate the temp list and using the index property, would it be easier, and clearer, to use a shuffled stride to assign a new value  to numbers and not use temp and index at all?

I see you're also comparing a lot of things with 'count.'  Is it not better to use the .isEmpty method call?

Sorry, I don't mean to hijack your work.  I found this looking for something else bingo related and thought I would comment based on what I see.